refactor(e2e): split actions.rs into submodule
